test: prove Android routing with third-party traffic
