+++ kmp_module/app/src/androidDebug/kotlin/com/dobby/feature/vpn_service/VpnConsentTestActivity.kt
@@ -3,9 +3,6 @@ package com.dobby.feature.vpn_service
 import android.app.Activity
 import android.net.VpnService
 import android.os.Bundle
-import java.net.DatagramPacket
-import java.net.DatagramSocket
-import java.net.InetAddress
 
 /**
  * Debug-only foreground host for instrumentation that exercises Android's real VPN consent UI.
@@ -14,23 +11,6 @@ import java.net.InetAddress
 class VpnConsentTestActivity : Activity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        if (intent.action == ACTION_SEND_DOCUMENTATION_PACKET) {
-            Thread(
-                {
-                    Thread.sleep(ROUTE_SETTLE_MILLIS)
-                    DatagramSocket().use { socket ->
-                        val target = InetAddress.getByName(DOCUMENTATION_ROUTE_ADDRESS)
-                        repeat(PACKET_BURST_COUNT) {
-                            socket.send(DatagramPacket(byteArrayOf(0x44), 1, target, DOCUMENTATION_ROUTE_PORT))
-                            Thread.sleep(PACKET_INTERVAL_MILLIS)
-                        }
-                    }
-                    runOnUiThread(::finish)
-                },
-                "vpn-test-packet",
-            ).start()
-            return
-        }
         val consentIntent = VpnService.prepare(this)
         if (consentIntent == null) {
             finish()
@@ -45,15 +25,7 @@ class VpnConsentTestActivity : Activity() {
         if (requestCode == VPN_CONSENT_REQUEST) finish()
     }
 
-    companion object {
-        const val ACTION_SEND_DOCUMENTATION_PACKET =
-            "com.dobby.feature.vpn_service.action.SEND_DOCUMENTATION_PACKET"
-
+    private companion object {
         private const val VPN_CONSENT_REQUEST = 1
-        private const val DOCUMENTATION_ROUTE_ADDRESS = "192.0.2.1"
-        private const val DOCUMENTATION_ROUTE_PORT = 33_434
-        private const val ROUTE_SETTLE_MILLIS = 500L
-        private const val PACKET_INTERVAL_MILLIS = 200L
-        private const val PACKET_BURST_COUNT = 5
     }
 }
